Extract ModelLineInfoMapBuilder as a standalone, publicly testable unit
buildModelLineMap was internal on ResultsFulfillerAppRunner specifically so
tests could inject fakes for blob loading and VID index construction --
exactly the pattern project testing standards call out: do not expose
internal functionality just for testability, and consider whether it
needs to be a separate unit instead.

Moves the grouping/caching/TypeRegistry-scoping logic into its own
ModelLineInfoMapBuilder class with a public build() method, decoupled from
CLI flags via a plain ModelLineSource data class instead of the Picocli
ModelLineFlags type (which would have created a circular BUILD
dependency between the builder and the CLI runner). Its tests move to
ModelLineInfoMapBuilderTest.kt, calling the public API directly instead of
an internal test seam.

Also declares the protobuf-java dependency explicitly on both the new
target and the existing results_fulfiller_app_runner target, instead of
relying on it being pulled in transitively through generated proto
targets.
